feat: swagger implementation
